// webapp/src/@generated/schemas/dailyVisitorDtoGroup.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * Vol
 * Registration api
 * OpenAPI spec version: 0.1
 */

export type DailyVisitorDtoGroup = typeof DailyVisitorDtoGroup[keyof typeof DailyVisitorDtoGroup];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DailyVisitorDtoGroup = {
  Unknown: 'Unknown',
  Adults: 'Adults',
  Infants: 'Infants',
  Children: 'Children',
  Teenagers: 'Teenagers',
  OlderAdults: 'OlderAdults',
} as const;