// webapp/src/@generated/schemas/visitorDtoGender.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * Vol
 * Registration api
 * OpenAPI spec version: 0.1
 */

export type VisitorDtoGender = typeof VisitorDtoGender[keyof typeof VisitorDtoGender];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VisitorDtoGender = {
  Male: 'Male',
  Female: 'Female',
} as const;