// webapp/src/@generated/schemas/visitorDtoStatus.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * Vol
 * Registration api
 * OpenAPI spec version: 0.1
 */

export type VisitorDtoStatus = typeof VisitorDtoStatus[keyof typeof VisitorDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VisitorDtoStatus = {
  Local: 'Local',
  Refugee: 'Refugee',
} as const;